+++ docs/changelog.md
@@ -29,6 +29,10 @@
 
 ## 更新说明
 
+1.6.9 fix: fix stock_zh_a_spot_em interface
+
+    1. 修复 stock_zh_a_spot_em 接口，新增多个字段
+
 1.6.8 fix: fix futures_zh_spot interface
 
     1. 修复 futures_zh_spot 接口
@@ -716,6 +720,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.6.9 fix: fix stock_zh_a_spot_em interface
 
 1.6.8 fix: fix futures_zh_spot interface